Update E2E test to include check cold key swap
